<Chuong>: Admin bổ ssung function

+++ SourceCode/DogCatManagement/DogCatManagement/AdminHomePage.cs
@@ -12,6 +12,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement;
 
 namespace DogCatManagement
 {
@@ -22,7 +23,7 @@ namespace DogCatManagement
         private IReportRepository _reportRepository;
         private UserSession _userSession;
         public AdminHomePage(UserSession userSession)
-        {   
+        {
             InitializeComponent();
             // Khai bao Service
             _userRepository = new UserRepository();
@@ -34,7 +35,36 @@ namespace DogCatManagement
             combobox.Add("customer", "Khách hàng");
             combobox.Add("post", "Bài đăng");
             combobox.Add("report", "Báo cáo");
-            cbbDataAdmin.DataSource = new BindingSource(combobox,null);
+            cbbDataAdmin.DataSource = new BindingSource(combobox, null);
+            cbbDataAdmin.DropDownStyle = ComboBoxStyle.DropDownList;
+            cbbDataAdmin.DisplayMember = "Value";
+            cbbDataAdmin.ValueMember = "Key";
+
+            // ComboBox User 
+            Dictionary<int, string> userStatus = new Dictionary<int, string>();
+            userStatus.Add(10, "Tất cả");
+            userStatus.Add(0, "Khách hàng");
+            userStatus.Add(1, "Nhân Viên");
+            userStatus.Add(2, "Admin");
+            userStatus.Add(3, "Bị Khóa");
+            cbbFilterUser.DataSource = new BindingSource(userStatus, null);
+            cbbFilterUser.DropDownStyle = ComboBoxStyle.DropDownList;
+            cbbFilterUser.DisplayMember = "Value";
+            cbbFilterUser.ValueMember = "Key";
+
+            // ComboBox Post
+            Dictionary<string, string> postStatus = new Dictionary<String, string>();
+            postStatus.Add("All", "Tất cả");
+            postStatus.Add("Rejected", "Từ chối");
+            postStatus.Add("Pending", "Đang chờ được duyệt");
+            postStatus.Add("Published", "Đã được duyệt");
+            cbbFilterPost.DataSource = new BindingSource(postStatus, null);
+            cbbFilterPost.DropDownStyle = ComboBoxStyle.DropDownList;
+            cbbFilterPost.DisplayMember = "Value";
+            cbbFilterPost.ValueMember = "Key";
+
+            // Combox Post
+            cbbDataAdmin.DataSource = new BindingSource(combobox, null);
             cbbDataAdmin.DisplayMember = "Value";
             cbbDataAdmin.ValueMember = "Key";
             LoadCustomerData();
@@ -53,9 +83,14 @@ namespace DogCatManagement
         // Logout Click
         private void btnLogout_Click(object sender, EventArgs e)
         {
-            LoginForm loginForm = new LoginForm();
-            this.Hide();
-            loginForm.Show();
+            DialogResult result = MessageBox.Show("Bạn có chắc chắn muốn đăng xuất?", "Xác nhận đăng xuất", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+            if (result == DialogResult.Yes)
+            {
+                LoginForm form = new LoginForm();
+                form.Show();
+                this.Close();
+            }
         }
         // Load Customer Data
         public void LoadCustomerData()
@@ -69,13 +104,15 @@ namespace DogCatManagement
             switch (selectedcbb)
             {
                 case "post":
-                    dgvData.DataSource = dgvData.DataSource = _postRepository.GetAllPost().Select(c =>  new { c.UserId, c.PostId, c.Title, c.Content, c.PostDate, c.Status, c.User.UserName }).ToList();
+                    dgvData.DataSource = dgvData.DataSource = _postRepository.GetAllPost().Select(c => new { c.UserId, c.PostId, c.Title, c.Content, c.PostDate, c.Status, c.User.UserName }).ToList();
                     gbPost.Visible = true;
                     txtSearchPost.Visible = true;
                     txtReportSearch.Visible = false;
                     txtSearchUser.Visible = false;
                     gbCustomer.Visible = false;
                     gbReport.Visible = false;
+                    cbbFilterPost.Visible = true;
+                    cbbFilterUser.Visible = false;
                     break;
                 case "customer":
                     dgvData.DataSource = _userRepository.GetAllUserList();
@@ -85,15 +122,19 @@ namespace DogCatManagement
                     txtSearchPost.Visible = false;
                     txtReportSearch.Visible = false;
                     txtSearchUser.Visible = true;
+                    cbbFilterPost.Visible = false;
+                    cbbFilterUser.Visible = true;
                     break;
                 case "report":
-                    dgvData.DataSource = _reportRepository.GetAllReportList().Select(c => new {c.ReportId,c.UserId, c.PostId, c.ReportDate,c.ReportReason,c.User.UserName, c.Post.Title}).ToList();
-                    gbCustomer.Visible=false;
+                    dgvData.DataSource = _reportRepository.GetAllReportList().Where(m => m.Post.Status.Equals("Published")).Select(c => new { c.ReportId, c.UserId, c.PostId, c.ReportDate, c.ReportReason, c.User.UserName, c.Post.Title }).ToList();
+                    gbCustomer.Visible = false;
                     gbPost.Visible = false;
                     gbReport.Visible = true;
                     txtSearchPost.Visible = false;
                     txtReportSearch.Visible = true;
                     txtSearchUser.Visible = false;
+                    cbbFilterPost.Visible = false;
+                    cbbFilterUser.Visible = false;
                     break;
             }
         }
@@ -103,12 +144,12 @@ namespace DogCatManagement
         {
             Post post = new Post();
 
-                post.Title = "Nhập tiêu đề ở đây...";
-                post.Status = "Pending";
-                post.Content = "Nhập nội dung ở đây...";
-                post.PostDate = DateTime.Now;
-                post.UserId = _userSession.UserId;
-                return post;
+            post.Title = "Nhập tiêu đề ở đây...";
+            post.Status = "Pending";
+            post.Content = "Nhập nội dung ở đây...";
+            post.PostDate = DateTime.Now;
+            post.UserId = _userSession.UserId;
+            return post;
         }
         // Get Post From Data
         private Post getPostFromDataGirdView()
@@ -129,24 +170,42 @@ namespace DogCatManagement
         // Yêu cầu Post
         private void btnPostUpdate_Click(object sender, EventArgs e)
         {
-            Post updatepost = new Post();
-            updatepost = getPostFromDataGirdView();
-            this.Hide();
-            PostForm postform = new PostForm(updatepost,_userSession.UserId);
-            postform.Show();
+            if (dgvData.SelectedRows.Count <= 0)
+            {
+                MessageBox.Show("Hiện tại không ghi nhận bất cứ một bài đăng nào");
+            }
+            else
+            {
+                Post updatepost = new Post();
+                updatepost = getPostFromDataGirdView();
+                if (updatepost != null)
+                {
+                    if (updatepost.Status.Equals("Rejected"))
+                    {
+                        MessageBox.Show("This post has been rejected!!", "Message", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    else
+                    {
+                        this.Hide();
+                        PostForm postform = new PostForm(updatepost, _userSession.UserId);
+                        postform.Show();
+                    }
+                }
+            }
         }
         // Add Post
         private void btnPostAdd_Click(object sender, EventArgs e)
         {
             Post emptyPost = getEmptyPost();
-            PostForm postForm = new PostForm(emptyPost,_userSession.UserId);
+            PostForm postForm = new PostForm(emptyPost, _userSession.UserId);
             this.Hide();
             postForm.Show();
         }
         // Search Post
         private void txtSearchPost_TextChanged(object sender, EventArgs e)
         {
-
+            string keyword = txtSearchPost.Text.ToUpper();
+            dgvData.DataSource = _postRepository.GetAllPost().Where(m => m.User.UserName.ToUpper().Contains(keyword)).Select(c => new { c.UserId, c.PostId, c.Title, c.Content, c.PostDate, c.Status, c.User.UserName }).ToList();
         }
         // PostDelete   
         private void btnPostDelete_Click(object sender, EventArgs e)
@@ -158,6 +217,7 @@ namespace DogCatManagement
         private Report getReportFromDataGirdView()
         {
             Report report = new Report();
+
             if (dgvData.SelectedRows.Count > 0)
             {
                 DataGridViewRow selectedRow = dgvData.SelectedRows[0];
@@ -172,13 +232,20 @@ namespace DogCatManagement
         // View Detail Report
         private void btn_Click(object sender, EventArgs e)
         {
-            Report report = new Report();
-            report = getReportFromDataGirdView();
-            if (report != null)
+            if (dgvData.SelectedRows.Count == 0)
             {
-                ReportForm form = new ReportForm(report,_userSession);
-                this.Hide();
-                form.Show();
+                MessageBox.Show("Hiện tại không ghi nhận bất cứ phản hồi nào");
+            }
+            else if (dgvData.SelectedRows.Count > 0)
+            {
+                Report report = new Report();
+                report = getReportFromDataGirdView();
+                if (report != null)
+                {
+                    ReportForm form = new ReportForm(report, _userSession);
+                    this.Hide();
+                    form.Show();
+                }
             }
         }
         private void btnReportDelete_Click(object sender, EventArgs e)
@@ -186,18 +253,20 @@ namespace DogCatManagement
 
         }
 
+        // Search User 
         private void txtSearchAdmin_TextChanged(object sender, EventArgs e)
         {
-
+            string keyword = txtSearchUser.Text.ToUpper();
+            dgvData.DataSource = _userRepository.GetAllUserList().Where(k => k.Email.ToUpper().Contains(keyword)).ToList();
         }
         // Customer Area
         // Button Add User
         private void btnAdd_Click(object sender, EventArgs e)
         {
-            UserDTO user = getEmptyUser(); 
+            UserDTO user = getEmptyUser();
             CustomerForm frm = new CustomerForm(user, _userSession);
             this.Hide();
-            frm.Show();  
+            frm.Show();
         }
         // Button Update User
         private void btnUpdate_Click(object sender, EventArgs e)
@@ -218,12 +287,19 @@ namespace DogCatManagement
                 user.Email = selectedRow.Cells["Email"].Value.ToString();
                 user.UserName = selectedRow.Cells["Username"].Value.ToString();
                 user.Password = selectedRow.Cells["Password"].Value.ToString();
-                user.RoleId =  Int32.Parse(selectedRow.Cells["RoleId"].Value.ToString());
-                if (MessageBox.Show("Bạn có chắc muốn khóa người dùng " + user.UserName + " không?", "Thông báo", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == DialogResult.OK)
+                user.RoleId = Int32.Parse(selectedRow.Cells["RoleId"].Value.ToString());
+                if (user.RoleId == 3)
                 {
-                    _userRepository.DeleteUser(user);
-                    MessageBox.Show("Bạn đã khóa thành công người dùng này");
-                    dgvData.DataSource = _userRepository.GetAllUserList();
+                    MessageBox.Show("Hiện tại người dùng đã bị khóa, bạn có thể vào phần chỉnh sửa để chỉnh sửa giới hạn truy cập của user này!!");
+                }
+                else
+                {
+                    if (MessageBox.Show("Bạn có chắc muốn khóa người dùng " + user.UserName + " không?", "Thông báo", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == DialogResult.OK)
+                    {
+                        _userRepository.DeleteUser(user);
+                        MessageBox.Show("Bạn đã khóa thành công người dùng này");
+                        dgvData.DataSource = _userRepository.GetAllUserList();
+                    }
                 }
             }
         }
@@ -256,9 +332,71 @@ namespace DogCatManagement
             }
             return user;
         }
+        //
         private void txtReportSearch_TextChanged(object sender, EventArgs e)
         {
 
+        }
+
+        private void dgvData_CellFormatting(object sender, DataGridViewCellFormattingEventArgs e)
+        {
+            if (dgvData.Columns[e.ColumnIndex].Name == "Password")
+            {
+                if (e.Value != null)
+                {
+                    string password = e.Value.ToString();
+                    e.Value = new string('*', password.Length);
+                }
+            }
+        }
+
+        private void gbCustomer_Enter(object sender, EventArgs e)
+        {
+
+        }
+
+        // Filter User By Role ID
+        private void cbbFilterUser_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            int value = ((KeyValuePair<int, string>)cbbFilterUser.SelectedItem).Key;
+            switch (value)
+            {
+                default:
+                    dgvData.DataSource = _userRepository.GetAllUserList();
+                    break;
+                case 0:
+                    dgvData.DataSource = _userRepository.GetAllUserList().Where(m => m.RoleId == 0).ToList();
+                    break;
+                case 1:
+                    dgvData.DataSource = _userRepository.GetAllUserList().Where(m => m.RoleId == 1).ToList();
+                    break;
+                case 2:
+                    dgvData.DataSource = _userRepository.GetAllUserList().Where(m => m.RoleId == 2).ToList();
+                    break;
+                case 3:
+                    dgvData.DataSource = _userRepository.GetAllUserList().Where(m => m.RoleId == 3).ToList();
+                    break;
+            }
+        }
+        // Filter Post By Status
+        private void cbbFilterPost_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            string value = cbbFilterPost.SelectedValue.ToString();
+            switch (value)
+            {
+                case "All":
+                    dgvData.DataSource = dgvData.DataSource = _postRepository.GetAllPost().Select(c => new { c.UserId, c.PostId, c.Title, c.Content, c.PostDate, c.Status, c.User.UserName }).ToList();
+                    break;
+                case "Rejected":
+                    dgvData.DataSource = dgvData.DataSource = _postRepository.GetAllPost().Where(p => p.Status.Equals("Rejected")).Select(c => new { c.UserId, c.PostId, c.Title, c.Content, c.PostDate, c.Status, c.User.UserName }).ToList();
+                    break;
+                case "Pending":
+                    dgvData.DataSource = dgvData.DataSource = _postRepository.GetAllPost().Where(p => p.Status.Equals("Pending")).Select(c => new { c.UserId, c.PostId, c.Title, c.Content, c.PostDate, c.Status, c.User.UserName }).ToList();
+                    break;
+                case "Published":
+                    dgvData.DataSource = dgvData.DataSource = _postRepository.GetAllPost().Where(p => p.Status.Equals("Published")).Select(c => new { c.UserId, c.PostId, c.Title, c.Content, c.PostDate, c.Status, c.User.UserName }).ToList();
+                    break;
+            }
         }
     }
 }
